+++ src/myarchive/db/shotwell/tables.py
